% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesTLL.R
\name{real_TLL}
\alias{real_TLL}
\title{Bayesian and Classical Estimates of Topp-leone distribution
Bayesian and Classical Estimates of TLL distribution uniform prior distribution for scale parameter and gamma distribution for shape parameter under Linex loss function, Square error loss function and Entropy loss function.}
\usage{
real_TLL(x, k, p, q, k1)
}
\arguments{
\item{x}{Numeric vector.}

\item{k}{Linex loss function parameter \code{numeric}}

\item{p}{shape parameter of a gamma distribution \code{numeric}
#'@param alpha1 shape \code{numeric}}

\item{k1}{Entropy loss function \code{numeric}}

\item{theta1}{rate \code{numeric}}
}
\value{
Bayesian and ML estimates on real dataset
}
\description{
Bayesian and Classical Estimates of Topp-leone distribution
Bayesian and Classical Estimates of TLL distribution uniform prior distribution for scale parameter and gamma distribution for shape parameter under Linex loss function, Square error loss function and Entropy loss function.
}
\examples{
x = Rtll(n=50, alpha=1,theta=1)
real_TLL(x, k=1,p=1,q=1,k1=1)
}
