% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesTLL.R
\name{Rtll}
\alias{Rtll}
\title{Simulate data from TLL distribution}
\usage{
Rtll(n, alpha, theta)
}
\arguments{
\item{n}{number of sample \code{Numeric}}

\item{alpha}{which is the shape parameter  \code{numeric}}

\item{theta}{which is the rate  rate\code{numeric}}
}
\description{
Simulate data from TLL distribution
}
\examples{
x = Rtll(n=50, alpha=2,theta=1)

}
