% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesTLL.R
\name{sim_TLL}
\alias{sim_TLL}
\title{Bayesian and Classical Estimates of Topp-leone distribution
Bayesian and Classical Estimates of TLL distribution uniform prior distribution for scale parameter and gamma distribution for shape parameter under Linex loss function, Square error loss function and Entropy loss function.}
\usage{
sim_TLL(n, r, alpha1, theta1, k, p, q, k1)
}
\arguments{
\item{n}{scale parameter of a gamma distrbution \code{numeric}}

\item{r}{number of iteration \code{numeric}}

\item{k}{Linex loss function parameter \code{numeric}}

\item{p}{shape parameter of a gamma distribution \code{numeric}}

\item{k1}{Entropy loss function \code{numeric}}

\item{x}{Numeric vector.}
}
\value{
Bayesian and ML estimates
}
\description{
Bayesian and Classical Estimates of Topp-leone distribution
Bayesian and Classical Estimates of TLL distribution uniform prior distribution for scale parameter and gamma distribution for shape parameter under Linex loss function, Square error loss function and Entropy loss function.
}
\examples{

sim_TLL(n=500, r =1000, alpha=1,theta=2, k=1,p=2,q=1,k1=1)

}
